Squashed commit of the following:
commit 85a2c907503ff83e21c59a03458a2c33b6ce4cc9
Author: d0UBleW <66501624+d0UBleW@users.noreply.github.com>
Date:   Sun Apr 17 20:00:14 2022 +0800

    fix(sort): initialized function pointer

+++ DSTR-Assignment-Array/sort.cpp
@@ -29,7 +29,7 @@ int CompareString(std::string s1, std::string s2) {
         if (s1[i] < s2[i]) return -1;
         if (s1[i] > s2[i]) return 1;
     }
-    return s1.length() - s2.length();
+    return (int) s1.length() - (int) s2.length();
 }
 
 int CompareTutorID(Tutor &t1, Tutor &t2) {
